+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-* Nothing
+### Fixed
+
+* Fixed parsing of empty singular relationships.
 
 ## [1.0.0-beta.2] - 2019-09-20